+++ src/test/java/net/kaczmarzyk/utils/interceptor/InterceptedStatementsAssert.java
@@ -17,6 +17,7 @@ package net.kaczmarzyk.utils.interceptor;
 
 import java.util.ArrayList;
 import java.util.List;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -116,5 +117,16 @@ public class InterceptedStatementsAssert {
 
 		return count;
 
+	}
+  
+  public InterceptedStatementsAssert hasSingleSelectWithNumberOfJoins(int expectedNumberOfJoins) {
+		assertThat(logs.size()).isEqualTo(1);
+
+		Integer numberOfJoins = countNumberOfSqlClauseInStatement(logs.get(0), "join");
+
+		assertThat(numberOfJoins)
+				.isEqualTo(expectedNumberOfJoins);
+
+		return this;
 	}
 }